Add TopicModelling class for training and inference based on the prototype solution
Notebook 01 is the report for the prototyping of the solution on the topic
modelling problem. I also modified the requirements.txt file to ensure that
the environment is the same as when run on Google Colab.

I also deleted some unnecessary files from the Cookiecutter Data Science template.
And added an initial docker-compose file to boot up MongoDB database for
future backend development.